polish history code format

+++ paddle/fluid/operators/sequence_ops/sequence_erase_op.h
@@ -35,12 +35,12 @@ class SequenceEraseKernel : public framework::OpKernel<T> {
         platform::errors::InvalidArgument("Input(X) Tensor of SequenceEraseOp "
                                           "does not contain LoD information."));
     PADDLE_ENFORCE_EQ(lod[lod.size() - 1].back(),
-                      (size_t)in->numel(),
+                      static_cast<size_t>(in->numel()),
                       platform::errors::InvalidArgument(
                           "The actual input size %d mismatches with the LoD "
                           "information size %d.",
                           lod[lod.size() - 1].back(),
-                          (size_t)in->numel()));
+                          in->numel()));
     auto tokens = ctx.Attr<std::vector<int>>("tokens");
     auto in_len = in->numel();
     auto in_dat = in->data<T>();
